Adding wPLI to the connectivity measures, for #67 (#121)
* Adding wPLI to the connectivity measures

A new function "_multiply_conjugate_time" was added. This function is a slight modification of the pre-existing "_multiply_conjugate". 
"_multiply_conjugate_time" computes the product of a complex array and its conjugate, but (unlike "_multiply_conjugate") doesnt collapse the last dimmension (i.e., doesnt sum the results of the multiplication, preserving the product values for each time point - which are needed to compute the wPLI).

wPLI was added to the pre-existing "compute_sync" function.

* Adding PLI to the connectivity measures

Adds the PLI to the pre-existing "compute_sync" function.