Fix bug when 2 signals in a pair are not perfectly aligned for wavelet coherence (#232)
Add length of signal to the wavelet caching key.

There are cases where 2 signals in a pair are not exactly the same
length (one sample of difference). In these cases the longest signal is
cropped. But when computing intra, we reuse the same signal (not from
the pair), so we should load a cropped cached signal.